requirements.txt
Flask==1.1.4
markupsafe==1.1.1
Flask_Login==0.6.3
python-dotenv==1.0.1
pypinyin==0.51.0
mysql
